source-google-ads/pyproject.toml
[tool.poetry]
name = "source_google_ads"
version = "0.1.0"
description = ""
authors = ["Jonathan Wihl <jonathan@estuary.dev>"]

[tool.poetry.dependencies]
airbyte-cdk = "^0.52"
estuary-cdk = {path="../estuary-cdk", develop = true}
freezegun = "^1.4.0"
google-ads = "^28.0.0"
protobuf = ">=5.29.5"
pydantic = "==1.10.14"
python = ">=3.11,<3.13"
types-requests = "^2.31"
urllib3 = ">=2.5.0"
requests = ">=2.32.4"
aiohttp = ">=3.10.11"
pycares = ">=4.8.1"

[tool.poetry.group.dev.dependencies]
debugpy = "^1.8.0"
pytest = "^7.4.3"
pytest-insta = "^0.2.0"
requests-mock = "^1.11.0"
pytest-mock = "^3.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

